Rebase mode: Make `rebase-mode-kill-line` comment a line instead of deleting it
It's disconcerting to see a line of text disappear with a single
keystroke with no indication of how to get it back.  The intention of
this change is to make `k` more similar to the other keys that change
the action of a line.

At some point in the future we may want to provide a keybinding that
really does "kill" a line.  (In fact, that function should probably
end up with the name "rebase-mode-kill-line".)  For now I think this
is the more friendly functionality for the `k` key.