Rebase mode: Make it easy to kill/unkill individual lines
Add #s to the regular expressions so that killed lines are recognized
as action/exec lines.  That makes it easy to move them around like
other lines, and you can use the normal commands for changing an
action line to change it into an un-killed action.

Modify `rebase-mode-exec-line` so that it can be used to un-kill an
exec line like the action commands can.