Fix error with "server-edit" and buffer-read-only in rebase-mode
It was previously impossible to actually use rebase-mode because you
could not save the buffer after changing it. Pressing "q" or "C-c C-c"
just caused an error about the buffer being read only. This change
adds "rebase-mode-server-edit", which uses the same workaround as
"rebase-mode-undo", and binds it to the appropriate keys in
"rebase-mode-map".