git-rebase-mode: define actions explicitly
Functions and especially commands should always be defined
explicitly using `defun' or a custom "define" macro.  This
allows e.g. `find-function' to locate the definition.  In
this case it also help people who want to customize the
keymap.

We are getting all the benefits for free.  It turned out
being explicit does not require a single additional line
of code.  We should still generate doc-strings though.