merge fns into git-commit-insert-header-as-self
Merge functionality provided functions that are exclusively used by
`git-commit-insert-header-as-self' into that function instead.

Splitting stuff into small functions is nice and all but when these new
functions are only used in one place and the also have to be documented
it makes little sense.  Especially if that makes understanding the only
caller *harder* (not much but still) and requires ~12 additional lines
of code and ~25 additional lines of comments.