`git-rebase-mode-map' bind "p" to `git-rebase-pick'
When rebasing interactively using git-rebase(1), the in-buffer comments
state that "p" means to "pick" the commit at the current line, but in
`git-rebase-mode', "p" means "move forward one line", so we need to use
another key ("c") to pick a commit.

Since we like being consistent with Git (see e.g. magit/magit#783),
bind "p" to `git-rebase-pick', and remove "n" and "p".  We're all used
to moving around with "C-n" and "C-p" anyway, aren't we?

Signed-off-by: Pieter Praet <pieter@praet.org>