`gitattributes-mode-{forward,backward}-field': DWIM
`gitattributes-mode-{forward,backward}-field' searched a single space
char forward/backward, so when the attributes were aligned with multiple
spaces (see e.g. the examples in the gitattributes(5) man page), we moved
forward/backward a single space at a time instead of from field to field.

By matching a single space followed by the empty string at the beginning
of a word (optionally preceded by an exclamation point or dash), it does
what it's supposed to do a bit better, but it definitely isn't perfect yet.

Signed-off-by: Pieter Praet <pieter@praet.org>