git-commit-mode.el: fix customizable max summary length
Customizing the max summary length (using `customize-variable' or `setq')
didn't have any effect since `git-commit-summary-regexp' was set using a
function in `git-commit-max-summary-line-length's ":initialize" keyword,
which runs when the `defcustom' is *evaluated* (with its default value),
instead of when it's *set* (with the custom value).

Using the ":set" keyword would make it work when using `customize-variable'
or starting a fresh Emacs session, but not when using plain `setq'.

Let's keep it simple and just generate the commit summary regexp on the fly,
making it work in all cases...

Also, fix the commentary section in the header.

See commit 246ccaf2.

Signed-off-by: Pieter Praet <pieter@praet.org>