tell newcomment.el # is only comment at line start
Otherwise we get things like this:

A typical commit message referencing issue #xxx that is long enough to
                                           #need filling,

This is not related to fontifying of comments which already applies only
to lines beginning with #. It's also not related to the syntax-table,
middle of line #s still have comment syntax, but apparently this has no
effect.

Fixes #67.