git-rebase-description-face: inherit from default
Set the default value to nil, i.e. inherit from `default'.  Previously
this did inherit from `font-lock-comment-face', which isn't so good
because it is for the description of commits that are *not* commented
out, *and* is other text in the buffer which are comments and also use
the respective face.