git-rebase-show-commit: don't bind default-directory
Do not let-bind `default-directory' around the call to
`magit-show-commit'.  That function works in the gitdir without any
problems, so changing to the worktree is not actually necessary.

However `magit-show-commit' causes `magit-repository-buffers'
to be called.  That function then uses `magit-inside-worktree-p'
to determine whether it should offer to save a file; and the
unnecessary let-binding removed here, tricked it into believing
`.git/rebase-merge/git-rebase-todo` is not actually in the gitdir.