magit-rebase-exec: change semantics
When there is a command on the current line then always edit it.
Even when it is commented out, like before that is uncommented
but the prompting also happens.

A prefix argument now forces prompting for a new command, instead
of editing the command on the current line.  So the behavior of
the prefix argument is more or less reverses.

This also allows the removal of the utility functions
`git-rebase-looking-at-exec' and `git-rebase-looking-at-killed-exec'.