refactor rebase commands, slightly changing their behavior
Refactor the commands `git-rebase-move-line-up',
`git-rebase-move-line-down', `git-rebase-kill-line',
`git-rebase-rebase-exec', `git-rebase-show-commit' and as well as the
utility function `git-rebase-edit-line', which is used by many of the
other commands.

The behavior of these commands changes slightly, in that they no longer
try to avoid moving point to the beginning of the line.  Previously
some of them did, while others did not.  Now they consistently do what
I prefer (move to bol) and which is also easier to implement.

Thanks to that we can now remove the utility functions
`git-rebase-looking-at-action' and
`git-rebase-looking-at-action-or-exec', and replace the constants
`git-rebase-action-line-re', `git-rebase-exec-line-re' and
`git-rebase-dead-line-re' with just `git-rebase-line'.  Also define
some regexps directly in `git-rebase-mode-font-lock-keywords'.